+++ server/api/auth/index.js
@@ -1,0 +1,7 @@
+module.exports = [
+  require('./login'),
+  require('./recover'),
+  require('./register'),
+  require('./test'),
+  require('./verify')
+]
