+++ server/api/auth/register.js
@@ -1,10 +1,10 @@
 const uuidv4 = require('uuid/v4')
 const emailValidator = require('email-validator')
-const cache = require('../cache')
-const util = require('../util')
-const auth = require('../auth')
-const config = require('../../config')
-const { responses } = require('../responses')
+const cache = require('../../cache')
+const util = require('../../util')
+const auth = require('../../auth')
+const config = require('../../../config')
+const { responses } = require('../../responses')
 
 module.exports = {
   method: 'post',
