+++ server/api/index.js
@@ -6,18 +6,17 @@ const auth = require('../auth')
 const router = express.Router()
 
 const routes = [
-  require('./auth-login'),
-  require('./auth-recover'),
-  require('./auth-register'),
-  require('./auth-test'),
-  require('./auth-verify'),
   require('./leaderboard'),
   require('./submitflag'),
   require('./challenges'),
-  require('./users-me'),
-  require('./users-id'),
   require('./ctftime-leaderboard')
 ]
+  .concat(
+    require('./users')
+  )
+  .concat(
+    require('./auth')
+  )
 
 const validationParams = ['body', 'params', 'query']
 const routeValidators = routes.map((route) => {
