+++ server/api/users/index.js
@@ -1,0 +1,5 @@
+// Note that the order of exports is important.
+module.exports = [
+  require('./me'),
+  require('./id')
+]
